# Multithreaded speedups for CPU models.